--- a/Cargo.toml
+++ b/Cargo.toml
@@ -16,7 +16,7 @@ build = "build.rs"
 libc = "^0.2"
 cfg-if = "0.1"
 
-[target.'cfg(target_os = "macos")'.dependencies]
+[target.'cfg(any(target_os = "macos", target_os = "linux"))'.dependencies]
 rayon = "^1.0"
 
 [target.'cfg(windows)'.dependencies]