## Mac

Before:

```
test bench_new                ... bench:   6,277,338 ns/iter (+/- 776,626)
test bench_refresh_all        ... bench:   2,777,395 ns/iter (+/- 792,563)
test bench_refresh_disk_lists ... bench:     252,386 ns/iter (+/- 27,681)
test bench_refresh_disks      ... bench:         616 ns/iter (+/- 270)
test bench_refresh_network    ... bench:      44,931 ns/iter (+/- 10,870)
test bench_refresh_process    ... bench:       4,250 ns/iter (+/- 132)
test bench_refresh_processes  ... bench:   1,130,398 ns/iter (+/- 34,410)
test bench_refresh_system     ... bench:     701,725 ns/iter (+/- 167,934)
```

After:

```
test bench_new                ... bench:   3,998,801 ns/iter (+/- 233,505)
test bench_refresh_all        ... bench:   1,278,118 ns/iter (+/- 143,394)
test bench_refresh_disk_lists ... bench:     252,440 ns/iter (+/- 15,763)
test bench_refresh_disks      ... bench:         630 ns/iter (+/- 249)
test bench_refresh_network    ... bench:      44,671 ns/iter (+/- 1,680)
test bench_refresh_process    ... bench:       4,465 ns/iter (+/- 2,300)
test bench_refresh_processes  ... bench:     446,773 ns/iter (+/- 63,277)
test bench_refresh_system     ... bench:     562,777 ns/iter (+/- 81,104)
```

## Linux

Before:

```
test bench_new                ... bench:  56,906,064 ns/iter (+/- 2,197,002)
test bench_refresh_all        ... bench:  11,413,569 ns/iter (+/- 855,971)
test bench_refresh_disk_lists ... bench:      76,971 ns/iter (+/- 9,166)
test bench_refresh_disks      ... bench:      42,398 ns/iter (+/- 6,330)
test bench_refresh_network    ... bench:      14,234 ns/iter (+/- 1,325)
test bench_refresh_process    ... bench:         200 ns/iter (+/- 8)   <== this was a bug
test bench_refresh_processes  ... bench:  10,757,594 ns/iter (+/- 947,818)
test bench_refresh_system     ... bench:      86,459 ns/iter (+/- 7,633)
```

After:

```
test bench_new                ... bench:  12,685,631 ns/iter (+/- 4,123,168)
test bench_refresh_all        ... bench:   5,440,071 ns/iter (+/- 1,286,536)
test bench_refresh_disk_lists ... bench:      94,706 ns/iter (+/- 6,656)
test bench_refresh_disks      ... bench:      43,847 ns/iter (+/- 12,716)
test bench_refresh_network    ... bench:      14,833 ns/iter (+/- 1,227)
test bench_refresh_process    ... bench:       7,248 ns/iter (+/- 844)
test bench_refresh_processes  ... bench:   4,035,119 ns/iter (+/- 759,456)
test bench_refresh_system     ... bench:      86,421 ns/iter (+/- 5,897)
```

## Windows

Before:

```
test bench_new                ... bench: 591,527,220 ns/iter (+/- 263,498,337)
test bench_refresh_all        ... bench: 356,635,315 ns/iter (+/- 64,393,390)
test bench_refresh_disk_lists ... bench:     157,257 ns/iter (+/- 20,659)
test bench_refresh_disks      ... bench:      45,767 ns/iter (+/- 20,894)
test bench_refresh_network    ... bench:         487 ns/iter (+/- 39)
test bench_refresh_process    ... bench:       8,458 ns/iter (+/- 312)
test bench_refresh_processes  ... bench: 327,626,625 ns/iter (+/- 39,483,199)
test bench_refresh_system     ... bench:       1,336 ns/iter (+/- 155)
```

After:

```
test bench_new                ... bench: 310,174,084 ns/iter (+/- 182,279,794)
test bench_refresh_all        ... bench: 123,698,474 ns/iter (+/- 52,320,355)
test bench_refresh_disk_lists ... bench:     152,887 ns/iter (+/- 16,594)
test bench_refresh_disks      ... bench:      45,967 ns/iter (+/- 5,002)
test bench_refresh_network    ... bench:         485 ns/iter (+/- 16)
test bench_refresh_process    ... bench:       1,705 ns/iter (+/- 108)
test bench_refresh_processes  ... bench: 122,890,190 ns/iter (+/- 5,728,812)
test bench_refresh_system     ... bench:       1,304 ns/iter (+/- 94)
```